made exception handeling for every router v1.12.5

+++ app/routers/compiler_exam.py
@@ -3,7 +3,7 @@ from typing import Annotated, Optional
 from sqlalchemy.orm import Session
 
 
-from ..schemas.compiler_exam import ExamCreation, ExamCreationResponse, ExamUpdate
+from ..schemas.compiler_exam import ExamCreation, ExamCreationResponse, ExamUpdate, ExamQuestionResponse
 from ..models.compiler import CodingExam, Questions, CandidateExamResult
 from ..db import get_db
 
@@ -14,7 +14,33 @@ router = APIRouter(prefix="/exam", tags=["EXAM"])
 
 @router.post("/creation", response_model=ExamCreationResponse)
 async def exam_creation(
-    request_body:Annotated[ExamCreation, Body()],
+    request_body:Annotated[ExamCreation, Body(
+        example=
+            {
+                "title": "technical",
+                "description": "something....",
+                "collage":"SMVM",
+                "window_start": "2025-12-01T11:40:36.967Z",
+                "window_end": "2025-12-01T12:40:36.967Z",
+                "duration": 20,
+                "category": "texc",
+                "questions": {
+                    "1": {
+                    "question_bank_id": 2,
+                    "score": 10
+                    },
+                    "2": {
+                    "question_bank_id": 1,
+                    "score": 10
+                    },
+                    "3": {
+                    "question_bank_id": 3,
+                    "score": 50
+                    }
+                }
+            }
+        
+    )],
     db:Annotated[Session, Depends(get_db)],
 ):
     try:
@@ -25,7 +51,7 @@ async def exam_creation(
         ## i.e, check is quetion ids' are in questions table at server.
 
         ## get question ids' to check in Questions table at server
-        all_question_ids = [i.bank_question_id for i in request_body.questions.values()]
+        all_question_ids = [i.question_bank_id for i in request_body.questions.values()]
         db_question_ids = [i.question_id for i in db.query(Questions).filter(Questions.question_id.in_(all_question_ids)).all()]
 
         ## list to catch not founf questions at db
@@ -58,7 +84,34 @@ async def exam_creation(
 @router.put("/update", response_model=ExamCreationResponse)
 async def exam_update(
     exam_id:Annotated[int, Query()],
-    request_body:Annotated[ExamUpdate, Body()],
+    request_body:Annotated[ExamUpdate, Body(
+        example=
+            {
+                "title": "technical",
+                "description": "something....",
+                "collage":"SMVM",
+                "is_active":1,
+                "window_start": "2025-12-01T11:40:36.967Z",
+                "window_end": "2025-12-01T12:40:36.967Z",
+                "duration": 20,
+                "category": "texc",
+                "questions": {
+                    "1": {
+                    "question_bank_id": 2,
+                    "score": 10
+                    },
+                    "2": {
+                    "question_bank_id": 1,
+                    "score": 10
+                    },
+                    "3": {
+                    "question_bank_id": 3,
+                    "score": 50
+                    }
+                }
+            }
+        
+    )],
     db:Annotated[Session, Depends(get_db)],
 ):
     try:
@@ -78,6 +131,7 @@ async def exam_update(
                 detail=f"Empty data not updatable..."
             )
 
+
         ## if questions are updating
         if "questions" in request_body_dict:
             ## -----------------------------------------
@@ -87,7 +141,7 @@ async def exam_update(
             ## i.e, check is quetion ids' are in questions table at server.
 
             ## get question ids' to check in Questions table at server
-            all_question_ids = [i.bank_question_id for i in request_body.questions.values()]
+            all_question_ids = [i.question_bank_id for i in request_body.questions.values()]
             db_question_ids = [i.question_id for i in db.query(Questions).filter(Questions.question_id.in_(all_question_ids)).all()]
 
             ## list to catch not founf questions at db
@@ -120,6 +174,7 @@ async def get_exam(
     exam_id:Annotated[int|None, Query()] = None,
     title:Annotated[str|None, Query()] = None,
     category:Annotated[str|None, Query()] = None,
+    collage:Annotated[str|None, Query()] = None,
 ):
     try:
         ## create base query
@@ -132,6 +187,8 @@ async def get_exam(
             db_coding_exam_base = db_coding_exam_base.filter(CodingExam.title.ilike(f"%{title}%"))
         if category:
             db_coding_exam_base = db_coding_exam_base.filter(CodingExam.category.ilike(f"%{category}%"))
+        if collage:
+            db_coding_exam_base = db_coding_exam_base.filter(CodingExam.collage.ilike(f"%{collage}%"))
 
         ## get final result
         db_coding_exam_final = db_coding_exam_base.all()
@@ -143,7 +200,7 @@ async def get_exam(
 
 
 
-@router.delete("/delete", status_code=status.HTTP_204_NO_CONTENT)
+@router.delete("/delete", status_code=status.HTTP_200_OK)
 async def delete_exam(
     exam_id:Annotated[int, Query()],
     db:Annotated[Session, Depends(get_db)],
@@ -157,30 +214,39 @@ async def delete_exam(
                 detail=f"Exam ID: {exam_id} Not Found on Question table at server..."
             )
         
+        ## Avoid deleting exam when its ID at results table
+        if db.query(CandidateExamResult).filter(CandidateExamResult.exam_id == exam_id).first():
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail=f"Not able to delete the exam ID: {exam_id}, some results are based on this exam ID"
+            )
+        
         ## del the exam
         db.delete(db_coding_exam)
         db.commit()
-        return
+        return {"message":f"Exam ID: {exam_id} deleted successfully"}
     except Exception as e:
         raise e
     
 
-@router.get("/get/details", response_model=ExamCreationResponse)
+@router.get("/get/details", response_model=ExamQuestionResponse)
 async def get_exam(
     db:Annotated[Session, Depends(get_db)],
     exam_id:Annotated[int, Query()],
 ):
     try:
-        ## create base query
-        db_coding_exam = db.query(CodingExam)
-
-        ## apply the results
-        db_coding_exam = db_coding_exam.filter(CodingExam.id == exam_id).first()
+        ## validate the examId
+        db_coding_exam = db.query(CodingExam).filter(CodingExam.id == exam_id).first()
+        if not db_coding_exam:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"Exam ID: {exam_id} Not Found on Question table at server..."
+            )
         
 
         for exam_question, value in db_coding_exam.questions.items():
             ## get details from question bank
-            details = db.query(Questions).filter(Questions.question_id == value["bank_question_id"]).first()
+            details = db.query(Questions).filter(Questions.question_id == value["question_bank_id"]).first()
 
             ## push each bank question deatils into regarding exam question
             db_coding_exam.questions[exam_question]["details"] = details
@@ -209,34 +275,7 @@ def get_results(
 
         records = query.all()
  
-        if not records:
-            missing = []
-            if candidate_id:
-                missing.append(f"candidate_id={candidate_id}")
-            if exam_id is not None:
-                missing.append(f"exam_id={exam_id}")
- 
-            message = "No results found"
-            if missing:
-                message += f" for ({', '.join(missing)})"
- 
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=message
-            )
- 
-        formatted = [
-            {
-                "result_id": item.result_id,
-                "candidate_id": item.candidate_id,
-                "exam_id": item.exam_id,
-                "total_marks": item.total,
-                "scores": item.score,
-            }
-            for item in records
-        ]
- 
-        return {"count": len(formatted), "results": formatted}
+        return records
  
     except Exception as e:
         raise e
